pyproject.toml
[project]
name = "agent-paper-explainer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.13",
    "langchain-deepseek>=1.0.1",
    "langgraph>=1.1.3",
    "python-dotenv>=1.2.2",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "chromadb>=0.4.0",
    "streamlit>=1.28.0",
    "pypdf>=4.0.0",
    "openai>=1.0.0",
    "sentence-transformers>=2.2.0",
    "grobid-client-python>=0.0.5",
    "requests>=2.31.0",
    "lxml>=4.9.0",
    "docker>=6.0.0",
    "feedparser>=6.0.10",
    "networkx>=3.1",
    "matplotlib>=3.7.0",
    "plotly>=5.15.0",
]
